Rename like.js
